Updated UI

+++ compx374winform/Form1.cs
@@ -425,8 +425,12 @@ namespace compx374winform
 
         private async void ButtonDeleteModel_Click(object sender, EventArgs e)
         {
-            await trainingClient.DeleteModelAsync(modelId = listBoxModels.SelectedItem.ToString().Split(' ')[0]);
-            LoadModels();
+            DialogResult dialogResult = MessageBox.Show("Are you sure you want to delete this model?", "Warning", MessageBoxButtons.YesNo);
+            if (dialogResult == DialogResult.Yes)
+            {
+                await trainingClient.DeleteModelAsync(modelId = listBoxModels.SelectedItem.ToString().Split(' ')[0]);
+                LoadModels();
+            }
         }
 
         private void ApiKeysToolStripMenuItem_Click(object sender, EventArgs e)
@@ -445,6 +449,11 @@ namespace compx374winform
             {
                 apiKeysForm.Focus();
             }
+        }
+
+        private void ButtonNewModel_Click(object sender, EventArgs e)
+        {
+
         }
     }
 }
